+++ echoframework/src/test/kotlin/org/echo/mobile/framework/model/socketoperations/GetAccountHistorySocketOperationTest.kt
@@ -105,7 +105,7 @@ class GetAccountHistorySocketOperationTest {
                                             {
                                             "id": "1.6.506",
                                             "op": [
-                                                25,
+                                                31,
                                                 {
                                                   "fee": {
                                                     "amount": 20,